Add a TacticianTestCase
And prepare for more than just the original tests.
