+++ test/Fixtures/DummyHandler.php
@@ -10,9 +10,14 @@
  * @license     http://opensource.org/licenses/Apache-2.0 Apache-2.0
  */
 
-namespace PMG\Queue\Tactician\Fixtures;
+namespace PMG\Queue\Fixtures;
 
-class NotMessage
+class DummyHandler
 {
+    public $command;
 
+    public function handle($command)
+    {
+        $this->command = $command;
+    }
 }
